# .stryker-tmp/sandbox-bCteHS/docs/api-contract.yaml
openapi: 3.0.3
info:
  title: VibeX API Contract
  description: |
    VibeX 全栈项目 API 契约文档
    
    ## 版本历史
    - v1.1.0 (2026-03-28): 从 14 端点扩展至 90+ 端点，覆盖全部前端实际调用
    - v1.0.0 (初始): 基础 14 端点定义
    
    ## 路径约定
    - 所有路径相对于 `/api` 前缀（如 `/auth/login` → `/api/auth/login`）
    - 服务器基础路径: `/api`
    - 统一无 v1 前缀（v1 双写路由已废弃，标注 `deprecated: true`）
    
    ## 路由分类
    - **前端调用**: 前端 services/api/ 模块中实际调用
    - **后端独有**: `x-backend-only: true`，前端未调用但后端已实现
    - **废弃路由**: `deprecated: true`，v1 双写兼容层
    
    ## 16 个 Tag 分组
    | Tag | 前缀 | 说明 |
    |-----|------|------|
    | Auth | /auth | 认证 |
    | Users | /users | 用户 |
    | Projects | /projects | 项目管理 |
    | Requirements | /requirements | 需求 |
    | Clarifications | /clarifications | 澄清问答 |
    | DomainEntities | /domain-entities, /domains | 域实体 |
    | EntityRelations | /entity-relations | 实体关系 |
    | DDD | /ddd | DDD 限界上下文/域模型/业务流程 |
    | Design | /clarify, /domain, /flow, /pages, /prototype, /design | 设计流程 |
    | Flows | /flows | 流程图 |
    | PrototypeSnapshots | /prototype-snapshots | 原型快照 |
    | Agents | /agents | Agent |
    | Pages | /pages | 页面 |
    | Chat | /chat | AI 对话 |
    | Plan | /plan | 计划分析 |
    | Diagnosis | /diagnosis | 需求诊断 |
    | BackendOnly | 各类后端独有路由 | 未被前端调用的后端路由 |
    | Deprecated | /api/v1/* | v1 兼容层（废弃） |
  version: 1.1.0
  contact:
    name: VibeX Team
servers:
  - url: /api
    description: API 基础路径
...
paths:
...
  /flows:
    get:
      tags: [Flows]
      summary: 获取流程图列表
      operationId: getFlows
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  flows:
                    type: array
                    items:
                      $ref: '#/components/schemas/FlowData'
    post:
      tags: [Flows]
      summary: 创建流程图
      operationId: createFlow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId]
              properties:
                projectId:
                  type: string
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowData'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /flows/{flowId}:
    get:
      tags: [Flows]
      summary: 获取单个流程图
      operationId: getFlow
      parameters:
        - name: flowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowData'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Flows]
      summary: 更新流程图
      operationId: updateFlow
      parameters:
        - name: flowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlowDataUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowData'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Flows]
      summary: 删除流程图
      operationId: deleteFlow
      parameters:
        - name: flowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # =============================================================================
  # TAG: PROTOTYPE SNAPSHOTS
  # =============================================================================
  /prototype-snapshots:
    get:
      tags: [PrototypeSnapshots]
      summary: 获取原型快照列表
      operationId: getPrototypeSnapshots
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  prototypeSnapshots:
                    type: array
                    items:
                      $ref: '#/components/schemas/PrototypeSnapshot'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [PrototypeSnapshots]
      summary: 创建原型快照
      operationId: createPrototypeSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrototypeSnapshotCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  prototypeSnapshot:
                    $ref: '#/components/schemas/PrototypeSnapshot'
        '400':
          $ref: '#/components/responses/BadRequest'
  /prototype-snapshots/{snapshotId}:
    get:
      tags: [PrototypeSnapshots]
      summary: 获取单个原型快照
      operationId: getPrototypeSnapshot
      parameters:
        - name: snapshotId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  prototypeSnapshot:
                    $ref: '#/components/schemas/PrototypeSnapshot'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [PrototypeSnapshots]
      summary: 更新原型快照
      operationId: updatePrototypeSnapshot
      parameters:
        - name: snapshotId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrototypeSnapshotUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  prototypeSnapshot:
                    $ref: '#/components/schemas/PrototypeSnapshot'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [PrototypeSnapshots]
      summary: 删除原型快照
      operationId: deletePrototypeSnapshot
      parameters:
        - name: snapshotId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # =============================================================================
  # TAG: AGENTS
  # =============================================================================
  /agents:
    get:
      tags: [Agents]
      summary: 获取 Agent 列表
      operationId: getAgents
      parameters:
        - name: userId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
    post:
      tags: [Agents]
      summary: 创建 Agent
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent:
                    $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /pages/{pageId}:
    get:
      tags: [Pages]
      summary: 获取单个页面
      operationId: getPage
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    $ref: '#/components/schemas/Page'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Pages]
      summary: 更新页面
      operationId: updatePage
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PageUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    $ref: '#/components/schemas/Page'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Pages]
      summary: 删除页面
      operationId: deletePage
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/agents:
    get:
      tags: [Deprecated]
      summary: 获取 Agent 列表（v1 兼容，已废弃）
      deprecated: true
      operationId: getAgentsV1
      responses:
        '200':
          description: 成功（兼容模式）
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
    post:
      tags: [Deprecated]
      summary: 创建 Agent（v1 兼容，已废弃）
      deprecated: true
      operationId: createAgentV1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: 创建成功（兼容模式）
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent:
                    $ref: '#/components/schemas/Agent'
...
  /v1/pages/{pageId}:
    get:
      tags: [Deprecated]
      summary: 获取单个页面（v1 兼容，已废弃）
      deprecated: true
      operationId: getPageV1
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功（兼容模式）
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    $ref: '#/components/schemas/Page'
    put:
      tags: [Deprecated]
      summary: 更新页面（v1 兼容，已废弃）
      deprecated: true
      operationId: updatePageV1
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PageUpdate'
      responses:
        '200':
          description: 更新成功（兼容模式）
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    $ref: '#/components/schemas/Page'
    delete:
      tags: [Deprecated]
      summary: 删除页面（v1 兼容，已废弃）
      deprecated: true
      operationId: deletePageV1
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功（兼容模式）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /v1/flows/{flowId}:
    get:
      tags: [Deprecated]
      summary: 获取单个流程图（v1 兼容，已废弃）
      deprecated: true
      operationId: getFlowV1
      parameters:
        - name: flowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功（兼容模式）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowData'
    put:
      tags: [Deprecated]
      summary: 更新流程图（v1 兼容，已废弃）
      deprecated: true
      operationId: updateFlowV1
      parameters:
        - name: flowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlowDataUpdate'
      responses:
        '200':
          description: 更新成功（兼容模式）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowData'
    delete:
      tags: [Deprecated]
      summary: 删除流程图（v1 兼容，已废弃）
      deprecated: true
      operationId: deleteFlowV1
      parameters:
        - name: flowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功（兼容模式）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Project:
      type: object
      required:
        - id
        - name
        - userId
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        userId:
          type: string
          format: uuid
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        pages:
          type: array
          items:
            $ref: '#/components/schemas/Page'
...
    FlowData:
      type: object
      required:
        - id
        - nodes
        - edges
        - projectId
      properties:
        id:
          type: string
          format: uuid
        nodes:
          type: array
          items:
            type: object
        edges:
          type: array
          items:
            type: object
        projectId:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FlowDataUpdate:
      type: object
      properties:
        nodes:
          type: array
          items:
            type: object
        edges:
          type: array
          items:
            type: object
        name:
          type: string
          nullable: true

    # ===== 原型快照相关 =====
    PrototypeSnapshot:
      type: object
      required:
        - id
        - projectId
        - content
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        version:
          type: integer
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PrototypeSnapshotCreate:
      type: object
      required:
        - projectId
        - content
      properties:
        projectId:
          type: string
        name:
          type: string
        description:
          type: string
        content:
          type: string
    PrototypeSnapshotUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        content:
          type: string

    # ===== Agent 相关 =====
    Agent:
      type: object
      required:
        - id
        - name
        - prompt
        - userId
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prompt:
          type: string
        model:
          type: string
          default: abab6.5s-chat
        temperature:
          type: number
          minimum: 0
          maximum: 1
          default: 0.7
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentCreate:
      type: object
      required:
        - name
        - prompt
        - userId
      properties:
        name:
          type: string
        prompt:
          type: string
        model:
          type: string
          default: abab6.5s-chat
        temperature:
          type: number
          default: 0.7
        userId:
          type: string
...
    Page:
      type: object
      required:
        - id
        - name
        - projectId
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        content:
          type: string
          nullable: true
        projectId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        project:
          $ref: '#/components/schemas/Project'
...
    PageUpdate:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
          nullable: true

    # ===== 消息相关 =====
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: 错误消息
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'